+++ hw01_hello_otus/main.go
@@ -1,9 +1,11 @@
 package main
 
-func Reverse(s string) string {
-	r := []rune(s)
-	for i, j := 0, len(r)-1; i < len(r)/2; i, j = i+1, j-1 {
-		r[i], r[j] = r[j], r[i]
-	}
-	return string(r)
+import (
+	"fmt"
+
+	"golang.org/x/example/stringutil"
+)
+
+func main() {
+	fmt.Println(stringutil.Reverse("Hello, OTUS!"))
 }
